MIPS: OF: Rework the appended DTB handling to keep the PROM arguments
Appended DTB are detected during the early init, if found the firmware
arguments are overwriten to pass the DTB using the UHI boot protocol,
which destroy the command line passed by the PROM.

To fix this issue move the check for appended DTB after the PROM init.
This allow to read the command line passed by the PROM before
overwriting the firmware arguments to pass the appended DTB.

For raw boot the code to detect a DTB just has to be moved from head.S
to setup.c. For zboot the DTB is simply copied after the decompressed
kernel, for this we need the size of the decompressed kernel as the
decompress API doesn't provides it.

Signed-off-by: Alban Bedel <albeu@free.fr>